+++ crates/pipeline-types/src/transport/delta_table.rs
@@ -14,7 +14,7 @@ pub struct DeltaTableWriterConfig {
     /// * [Azure options](https://docs.rs/object_store/latest/object_store/azure/enum.AzureConfigKey.html)
     /// * [Amazon S3 options](https://docs.rs/object_store/latest/object_store/aws/enum.AmazonS3ConfigKey.html)
     /// * [Google Cloud Storage options](https://docs.rs/object_store/latest/object_store/gcp/enum.GoogleConfigKey.html)
-    #[serde(default)]
+    #[serde(flatten)]
     pub object_store_config: HashMap<String, String>,
     /// A bound on the size in bytes of a single Parquet file.
     pub max_parquet_file_size: Option<usize>,