+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.14.0] - 2024-04-17
+
 ### Added
 
 - [SQL] `ARG_MIN`, `ARG_MAX` aggregation functions